requirements.txt
aiohttp==3.9.1
aiohttp-wsgi==0.10.0
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
dataclasses-json==0.6.3
deepgram-sdk==3.0.3
exceptiongroup==1.2.0
flask==3.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
python-dotenv==1.0.0
sniffio==1.3.0
typing-extensions==4.9.0
typing-inspect==0.9.0
verboselogs==1.7
websockets==12.0
werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
